// src/app/modules/rider/type.ts
import mongoose from "mongoose";

export type RiderDocument = mongoose.Document & {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  phoneNumber: string;
  otp: string;
  expired_at: Date;
  verifiedEmail: boolean;
  verifiedPhoneNumber: boolean;
  status: boolean;
  image?: string;
  address: string;
  location: {
    latitude: number;
    longitude: number;
  };
  isAvailable: string;

  deviceType: string;
  deviceToken: string;
  deviceName: string;

  comparePassword(candidatePassword: string): Promise<boolean>;
  generateJWT(): Promise<string>;
};

export type UserDataType = {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  phoneNumber: string;
};